requirements.txt
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.9.5
django-address==0.1.5
django-allauth==0.25.2
django-contrib-comments==1.7.1
django-conversation==1.4.6
django-datetime-widget==0.9.3
django-jsonfield==0.9.19
django-libs==1.67.4
django-model-utils==2.5
django-notify-x==0.1.6
django-oauth==1.1
django-staticfiles==1.2.1
django-widget-tweaks==1.4.1
gunicorn==19.4.5
jsonfield==1.0.3
oauthlib==1.1.1
Pillow==3.2.0
PyJWT==1.4.0
python-social-auth==0.2.19
python3-openid==3.0.10
pytz==2016.4
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
whitenoise==3.0
psycopg2==2.5.4